+++ src/server/live-reload/live-reload.js
@@ -49,7 +49,7 @@ LiveReload.prototype._onFileChanged = function (fileRelativePath, parentDir) {
     if (this._forcePrepare) {
         // Sometimes, especially on Windows, prepare will fail because the modified file is locked for a short duration
         // after modification, so we try to prepare twice.
-        propagateChangePromise = retryAsync(this._project.prepare.bind(this._project), 2)
+        propagateChangePromise = this._project.prepare()
             .then(function () {
                 return false;
             });
@@ -81,22 +81,6 @@ LiveReload.prototype._onFileChanged = function (fileRelativePath, parentDir) {
 
 function copyFile(src, dest) {
     return Q.nfcall(ncp, src, dest);
-}
-
-function retryAsync(promiseFunc, maxTries, delay, iteration) {
-    delay = delay || 100;
-    iteration = iteration || 1;
-
-    return promiseFunc().catch(function (err) {
-        if (iteration < maxTries) {
-            return Q.delay(delay)
-                .then(function () {
-                    return retryAsync(promiseFunc, maxTries, delay, iteration + 1);
-                });
-        }
-
-        return Q.reject(err);
-    });
 }
 
 module.exports = LiveReload;
